refactor: remove the `since` param from GET /websocket_updates
I've realized it's probably pointless trying to sequence in time updates from
webhooks and the API requests done in the initial sync, since there's really no
guarantee that webhooks will arrive soon after the causing event happened. So I
should just update the local db without caring about when I receive the updates,
and if I want any sort of mechanism to prevent applying updates to the local db
because the updates are obsolete/outdated, that mechanism should definitely not
depend on when the updates were received according to Heimisch servers', or the
client's system times. (I'd imagine it'd be possible to do something like that based
on `updated_at` or `created_at` values on the updates, which would be values
from Github's actual dbs).